[Scheduler] Change RegisterFile container to SmallVector
Using SmallVector as the underling container for RegisterFiles allows
more flexibility when converting to OptSched registers. Being able to
dynamically resize RegisterFiles means that we aren’t required to know
the total number of registers for each type ahead of time. This is more
reliable when performing transformations of the register info from LLVM.