Add new heuristic scheduler type.
Add a sequential list scheduler which forces scheduling the instructions
in the source input order. This is useful if you have a specific
schedule you want to use as input to the enumerator.

Select scheduler in sched.ini like below:

// The scheduler used to find an initial feasible schedule.
// LIST: List scheduler
// SEQ: Sequential list scheduler
// ACO: Ant colony scheduler
LIST_SCHED_TYPE SEQ

The ACO scheduler is now selected with this option as well.