Add LLVM as a heuristic priority.
‘LLVM’ replaces ‘NID’ as a list scheduler heuristic priority that
represents the schedule order generated by LLVM’s converging machine
scheduler. Unlike the NID priority, LLVM can be added at any point in
the heuristic priority string. For example, ‘LUC_LLVM’ means LUC as the
first priority, and LLVM as the secondary priority. Whereas previously
in LUC_NID, NID refers to the source order of the instructions. Note
that NID as the sole heuristic no longer refers to the LLVM heuristic.
This makes old configuration files with this scheme incompatible.