Fixed bug where logger would output that shorter schedule is found.
Value returned from enumCrntSched_->GetCrntLngth() is 0 due to enumerator. Enumerator stops when current length is 0 and all possibilities are considered. Therefore, it should not be used as a comparison for schedule length found.

Use trgLngth instead because loop will terminate at the current target length if a solution is found.