[Two-Pass] Integrated two-pass implementation from AMDGPU to base OptSched scheduler. (#26)
This change allows the base OptSched scheduler to also use the two-pass approach that is implemented in the AMDGPU target scheduler.

There are a few changes that have been added:

    Added more settings that can be set for the two-pass scheduler.
    1.) Different timeouts for each pass
    2.) Different enumerator heuristic for each pass
    Moved region recording to base OptSched scheduler class to avoid code redundancy.
    Moved occupancy pass and ILP pass source code to base OptSched scheduler class since both use the same code (most of it was just changing the settings).

Changes were tested with plaidbench running both the original and the new changes 3 times each. Results are identical when taking into account random variation.

    Compile-time: +0.03% difference
    TIles per second: -0.12% difference
    Examples per second: +0.13% difference

Note: With these changes, to use the AMDGPU scheduler, USE_TWO_PASS must be enabled and the settings should be set with the new settings.