[Scripts] Added schedule length and occupancy scripts. Improved older scripts. (#31)
All scripts can now output a spreadsheet (enabled by default). This function can be disabled with the "--disable" option.

Additionally, scripts does not have to be in the same directory anymore. Use the "-i" option to pass in the path to a directory to process.

Additional flags has also been added. Use '--help' to display the list.

Fixed inconsistent use of single/double quotes. Fixed and added some more comments.