[Two-Pass] Fix schedule specific regions bug (#40)
Enabling the setting "SCHEDULE_SPECIFIC_REGIONS" causes regions that are not selected for scheduling to bypass the recording phase. This messes up the region numbering in the multiple passes approach and causes a bug where we would schedule incorrect or no regions at all. This patch makes it so that we record all regions even if they are going to be skipped when scheduling to keep proper numbering of the regions.