Create C++ CI GitHub Workflow / Action
Compiles and runs the unit tests on push and on pull request.

This uses ccache to greatly speed up the build. The first build may take
more than an hour, but subsequent builds take less than 10 minutes, as
most of the work is rebuilding LLVM, which we never change. Using ccache
allows us to just reuse the compilation results rather than recompile.