Add Superbuild (#54)
Add superbuild CMakeLists.txt

This uses forks of llvm-project and flang-llvm with our patches already
applied. ExternalProject's PATCH_COMMAND is flaky, sometimes requiring
removal of the build directory and a restart.

This superbuild script can also build flang. This option is OFF by
default.

By default, llvm + OptSched and the Flang builds are installed to a
subdirectory of the build directory. These can be changed to be system
directories if that is desired.

Each sub-project has configurable git repo urls and other requirements.
This allows this superbuild script to support differing configurations,
such as different llvm versions.

Implementation note: LLVM has an LLVM_EXTERNAL_PROJECT configuration
variable. This allows OptSched to be a subproject of LLVM without
requiring a nested directory structure.