[GT]: Relax recursive successor requirement
For the hypothetical schedule considering nodes A and B, where B appears
before A, given a register R used by B but not by A, there must be a
node C such that:

The old condition was: C in RcrsvScsr(B) (and hence C in RcrsvScsr(A)
because of an earlier RcrsvScsr(B) is a subset of RcrsvScsr(A)). We
can relax this requirement: C in RcrsvScsr(A).

The goal is to show that such a C must exist after A, given
the hypothetical: [..., B, ..., A, ...]. If C in RcsrvScsr(A) and C is
not A, then the hypothetical must actually be
[..., B, ..., A, ..., C, ...], meaning that we're good and can proceed.