Create GitHub Workflow to run clang-tidy-9
The clang-tidy check `readability-identifier-naming` produces thousands
of warnings, so we turn it off in the main step. It is still run, just
in its own noisy step so that we can still see the warnings and work
towards fixing them.

We use CMake to create a compile_commands.json to pass to clang-tidy.
CMake does so with `-DCMAKE_EXPORT_COMPILE_COMMANDS=ON`. So that this
compilation database is valid, we also build the targets that generate
header files via llvm-tblgen. To avoid having to build llvm-tblgen
ourselves, we use a prebuilt executable matching our LLVM version.

The AMDGPU parts require LLVM 7.x+, hence the usage of LLVM 7.x.