[Enumerator] Fix enumerator not properly updating priorities of instructions already in ready list. (#70)
The enumerator was not properly updating the priorities for instructions already in the ready list for dynamic heuristics like LUC. New instructions added to the ready list would have the proper priority but their priority does not get updated. This change updates the priorities for all instructions in the ready list when a dynamic heuristic is selected.

Regarding compile time statistics for OptSched, FFT with heuristic LUC_NID gained +2 optimal regions found in the second pass. PlaidML saw a -1 to -3 loss in the optimal regions found and +1 to 3 gain in timed out and not improved when using LUC as a first priority heuristic. I believe this is a reasonable tradeoff for the new changes.

No significant regressions were seen in execution-time performance. PlaidML reported a -0.26% to 0.15% difference and SHOC reported a 0.09% to 0.60% difference in geometric scores. Both were within random variation.