Rewrite CMake scripting to support standalone
These changes allow us to be built standalone, provided an existing LLVM
build can be found with find_package(LLVM).

This enables us to share a single build of LLVM rather than rebuilding
LLVM for each build configuration. It also makes it possible to use a
different configuration for our code as compared to LLVM, which may aide
debugging.

If unit tests are enabled, the LLVM package must have the
LLVM_INSTALL_UTILS option set so that we can find FileCheck and other
required tools.