GT: Only count definitely shortened live ranges
This fixes the live range covering implementation to only count
shortened live ranges which will definitely be shortened, rather than
might be shortened.

For graph transformations, since we do not know the schedule but are
working on the DDG, we must be absolutely certain that added edges won't
make anything worse. This means that we must err on the side of caution.
For finding lengthened ranges, any register whose live range might be
lengthened should be counted, as it is, because they are bad for us.
However, for finding shortened ranges, we must be completely sure, since
shortened ranges are good for us.