Enable asserts in CI build, improving CMake
Turns on `assert(...)`s, guarding with #ifs as appropriate.

Additionally, improve the CMake build script to allow configuration-time
setting of #defines (for IS_DEBUG and similar source configurations), as
well as libraries to link to (simplifying GPU builds).

The IS_DEBUG macro turns on more assertions, thus its inclusion in the
CI build.