Inform of auto-reformatter on failed clang-format
Creates a message in the Clang Format GitHub workflow's logs which
directs the PR author to the auto-reformat trigger.

Previously, people were not aware that they could run clang-format via
writing a comment with the trigger. Hopefully, this message will be
enough.

Having the GitHub bot respond with a comment informing the PR author of
this feature was considered, but it was dropped due to technical
limitations in GitHub. A PR from a fork has a different GITHUB_TOKEN
secret as the main repo, preventing the secret from being used to add a
comment.