Fix undefined behavior with llvm_unreachable() and require valid values for some settings instead of defaults. (#112)
We were using llvm_unreachable() incorrectly and it was causing undefined behavior. One example is when using an incorrect heuristic in combination with another heuristic such as NOD_CP, which should be NID_CP, it incorrectly gave LLVM_CP. This patch removes it in favor of Logger::Fatal() which will output an error message and cause the compiler to crash.

Some default values have also been removed and Logger::Fatal() has been put in place to ensure that a valid value is given. We do not want to automatically default to a set value and would rather instead have the program indicate that our value was incorrect.