Sequential list scheduler fix undefined behavior (#122)
Fix SequentialListScheduler::IsSequentialInstruction undefined behavior on Inst number 0

If Inst->GetNum() == 0 then crntSched_->GetSchedCycle() is passed a -1 which results in undefined behavior in GetSchedCycle when slotNum = slotForInst_[-1]; is executed. If a -1 is the value in slotForInst_[-1] the Sequential List Scheduler is caught in an infinite loop of stalls since instruction number 0 is never legal.