Add a JSON -> info log conversion script (#125)
Add a JSON -> info log conversion script

Translates logs with the EVENT: + JSON format to the old INFO: format.

This is intended to be used for script verification purposes. An updated
script's results can be compared with the un-updated script by
translating the JSON logs and passing it through the older script.