Fix missing recursive neighbor update
Fixes the updating of recursive neighbors to also update the current
nodes. The recursive neighbor accessors do not include the node itself,
that is A is not in RPred(A), and similarly for successors. Therefore,
UpdateRecursiveNeighbors(A, B) has to add B's recursive successors to A,
and vice versa for A and recursive predecessors.