Handle INFO logs and interrupted tags
From the PR review, it was pointed out that log tags (EVENT:) can be
interrupted by compilation commands, so we probably want to handle such
cases as well. It turns out that SPEC's compilation commands are
intended to always appear at the beginning of a line, so matching just
for the first character of a log tag would be unlikely to result in
false positives.

It was also pointed out that INFO logs can also be interrupted, so it
makes sense to fix them as well. This is especially true because of the
way that we split logfiles: we split off of
'INFO: ********** Opt Scheduling **********'.
If this string was interrupted by a compilation command, suddenly we'd
merge multiple blocks together in our analysis.