Add comment-based check triggering (#141)
Fixes the prior attempt, making it actually run the checks in response
to a comment request.

Comment triggers do not git checkout to the same state as commit
triggers, so there is some adaptation for comment-triggered checks by
using the `gh` commandline tool (GitHub CLI) to checkout the current PR
state.

Furthermore, comment triggers do not show up in the list of GitHub
checks for a PR, so the workflow also adds a comment to show the status
of checks once they have finished.

Note that the cache does not work for comment-triggered checks, so the
actual C++ build will take around 1.5 hours. This is a limitation in the
cache tool.