Add reduced latency pass patch (#150)
This patch adds a reduced latency pass that can be used with the two-pass algorithm. This new pass is intended to be done after the Occupancy pass and can be used with or in place of the ILP pass.

The new reduced latency pass can be enabled by modifying the `sched.ini` file and modifying the `PASS_ORDER` setting to include the value `3`. There are also settings to tune the pass such as `LATENCY_TARGETS`, `LATENCY_DIVISOR`, and `LATENCY_MINIMUM`.

Since this pass reduces latencies, we will not get accurate schedule length values so an additional pass can be enabled by setting `COMPILE_TIME_DATA_PASS YES`. This additional pass runs through the schedule again with normal latencies by using the sequential list scheduler. Note that this will increase compile-time due to adding an additional pass. If you do not care about the compile-time data of our scheduler during the reduced latency pass then this pass can be left disabled.

Additionally, in this patch range tightening and relaxed scheduling pruning during enumeration have also been disabled in the Occupancy pass. Since all latencies are 1 in the occupancy pass, we know the min/max length of schedules thus do not need length-related pruning. The lower bound length calculation done before enumerating has also been disabled in any regions with max latency <= 1 since in these cases the lower bound length should be the number of instructions in the region.