Fix previous commit not disabling range tightening.
Additionally, range tightening was previousl pruning stalls in the Occupancy pass. Since stalls do not make sense in the occupancy pass (all latencies are 1), this commit removes stalling in first pass. Moreover, it bypasses lower bound techniques for determining if a history node dominates an active node, which, again, are not providing any information in first pass when latencies are 1.