Refactor parse_args
Results in a simpler interface that is easier to understand and avoids
globals. In short, changes parsing via `type=` parameter into specifying
which args should be parsed to be parsed after the fact.

This also fixes the performance problem where `--help` after a logfile
would wait to parse the logfile before exiting with a usage message.