Fixes a bug in SLIL cost computation on the device (#161)
The issue was that after copying over the WeightedBitVector for liveRegs_ and livePhysRegs_ and allocating the device array for vctr_ for all of the WBV, since oneCnt_ was set to 0, WBV::Dev_Reset() was not setting the values of each vctr_ array to 0. This fix forces Dev_Reset to initialize all values in both vctr_ arrays to 0 by setting the first bit to true, and making oneCnt_ equal to 1.