Fix pheromone table lookups, heuristic usage, and chance of selecting semi-ready instructions (#2)
Previously, the pheromone table lookups were incorrect if the previous instruction selected, was a semi-ready instruction that we decided to wait to become fully latency-ready. This now functions correctly for the score lookups.

Considerations are now taken to decide whether or not to select semi-ready instructions.

Different heuristics are now used in different passes based off of what is in sched.ini.